refactor: rename all "full" sync to "slow" sync
