+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/UserSessionScopeCommon.kt
@@ -132,13 +132,13 @@ import com.wire.kalium.logic.functional.map
 import com.wire.kalium.logic.functional.onSuccess
 import com.wire.kalium.logic.sync.ObserveSyncStateUseCase
 import com.wire.kalium.logic.sync.SetConnectionPolicyUseCase
-import com.wire.kalium.logic.sync.SyncCriteriaProvider
-import com.wire.kalium.logic.sync.SyncCriteriaProviderImpl
+import com.wire.kalium.logic.sync.slow.SyncCriteriaProvider
+import com.wire.kalium.logic.sync.slow.SlowSyncCriteriaProviderImpl
 import com.wire.kalium.logic.sync.SyncManager
 import com.wire.kalium.logic.sync.SyncManagerImpl
-import com.wire.kalium.logic.sync.full.SlowSyncManager
-import com.wire.kalium.logic.sync.full.SlowSyncWorker
-import com.wire.kalium.logic.sync.full.SlowSyncWorkerImpl
+import com.wire.kalium.logic.sync.slow.SlowSyncManager
+import com.wire.kalium.logic.sync.slow.SlowSyncWorker
+import com.wire.kalium.logic.sync.slow.SlowSyncWorkerImpl
 import com.wire.kalium.logic.sync.incremental.EventGatherer
 import com.wire.kalium.logic.sync.incremental.EventGathererImpl
 import com.wire.kalium.logic.sync.incremental.EventProcessor
@@ -402,7 +402,7 @@ abstract class UserSessionScopeCommon internal constructor(
         )
 
     private val syncCriteriaProvider: SyncCriteriaProvider
-        get() = SyncCriteriaProviderImpl(clientRepository, logoutRepository)
+        get() = SlowSyncCriteriaProviderImpl(clientRepository, logoutRepository)
 
     val syncManager: SyncManager by lazy {
         SyncManagerImpl(
